Revert "Fix tests and move tests::util::TestHandler to connection::util::TestHandler"
This reverts commit 40cdec370f10bd6dcd5d60d84e84f5289a937cc3.
