+++ core/src/connection/pool.rs
@@ -969,12 +969,7 @@ impl<THandler: IntoConnectionHandler> PendingConnection<'_, THandler> {
     /// Aborts the connection attempt, closing the connection.
     pub fn abort(mut self) {
         if let Some(notifier) = self.entry.get_mut().abort_notifier.take() {
-            match notifier.send(task::PendingConnectionCommand::Abort) {
-                Ok(()) => {}
-                Err(e) => {
-                    log::debug!("Failed to pending connection: {:?}", e);
-                }
-            }
+            notifier.send(task::PendingConnectionCommand::Abort);
         }
     }
 }
@@ -1360,51 +1355,5 @@ impl<'a, K: 'a, V: 'a> EntryExt<'a, K, V> for hash_map::Entry<'a, K, V> {
             hash_map::Entry::Occupied(entry) => entry,
             hash_map::Entry::Vacant(_) => panic!("{}", msg),
         }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    use crate::{
-        connection::PendingOutboundConnectionError,
-        transport::dummy::DummyTransport,
-    };
-    use futures::{executor::block_on, future::poll_fn};
-
-    #[test]
-    fn aborting_pending_connection_surfaces_error() {
-        let mut pool = Pool::new(
-            PeerId::random(),
-            PoolConfig::default(),
-            ConnectionLimits::default(),
-        );
-        let target_peer = PeerId::random();
-
-        pool.add_outgoing(
-            DummyTransport::default(),
-            ["/ip4/127.0.0.1/tcp/1234".parse::<Multiaddr>().unwrap()].into_iter(),
-            Some(target_peer),
-            crate::connection::util::TestHandler(),
-        )
-        .expect("Failed to add an outgoing connection to test pool.");
-
-        // Disconnect from the peer, thus aborting all pending connections.
-        pool.disconnect(&target_peer);
-
-        // Poll the pool until the pending connection is aborted.
-        block_on(poll_fn(|cx| match pool.poll(cx) {
-            Poll::Pending => return Poll::Pending,
-            Poll::Ready(PoolEvent::PendingOutboundConnectionError {
-                error: PendingOutboundConnectionError::Aborted,
-                ..
-            }) => {
-                return Poll::Ready(());
-            }
-            Poll::Ready(_) => {
-                panic!("We should see an aborted error, nothing else.")
-            }
-        }));
     }
 }
