+++ core/tests/network_dial_error.rs
@@ -23,14 +23,14 @@ mod util;
 use futures::prelude::*;
 use libp2p_core::multiaddr::multiaddr;
 use libp2p_core::{
-    connection::{PendingConnectionError, util::TestHandler},
+    connection::PendingConnectionError,
     multiaddr::Protocol,
     network::{NetworkConfig, NetworkEvent},
     PeerId,
 };
 use rand::seq::SliceRandom;
 use std::{io, task::Poll};
-use util::test_network;
+use util::{test_network, TestHandler};
 
 #[test]
 fn deny_incoming_connec() {
